mobility_alpha: float type

+++ read_input/read_input.py
@@ -43,7 +43,7 @@ class ReadInput(object):
     self.domain = str(self.options.get('domain') or '2D')
     self.scheme = str(self.options.get('scheme') or 'HLP_2D')
 
-    self.mobility_alpha = int(self.options.get('mobility_alpha') or 1)
+    self.mobility_alpha = float(self.options.get('mobility_alpha') or 1)
     self.radius = float(self.options.get('radius') or 1.0)
     self.intrinsic_linear_velocity = float(self.options.get('intrinsic_linear_velocity') or 1.0)
     self.intrinsic_angular_velocity = float(self.options.get('intrinsic_angular_velocity') or 1.0)
